requirements.txt
aiodns
aiohttp
asynctest
Celery
coverage
factory_boy
flake8
gitlint
ipdb
ipython
isort
pre-commit
pydocstyle
python-dotenv
redis
requests
responses
slackclient
sphinx
sphinx-rtd-theme
tornado
